Updated project files after move of include files. ExtremaFindingAlgorithms.h fixed a warning. (comparing unsigned int with int)
git-svn-id: https://www.h3d.org:8090/H3DUtil/trunk/H3DUtil@42 875a7764-6f23-0410-b93c-dbbdadb1ea73

+++ include/H3DUtil/ExtremaFindingAlgorithms.h
@@ -85,7 +85,7 @@ namespace H3DUtil {
 
         get_psum< InputType, OutputType >( psum, p, sum, mpts );
 
-        while( max_iterations < 0 || nr_iterations < max_iterations ) {
+        while( max_iterations < 0 || (int)nr_iterations < max_iterations ) {
           // find highest, lowest and next-highest point
           ilo = 0;
           ihi = y[1] > y[2] ? (inhi=2,1) : (inhi = 1,2);
